[refactor/oljs/Browser] move functions into own file
